chore: move Response addition to req version 8 from 9
